+++ EnhanceQoL/EnhanceQoL.lua
@@ -3551,7 +3551,10 @@ end
 local eventHandlers = {
 	["ACTIVE_PLAYER_SPECIALIZATION_CHANGED"] = function(arg1)
 		addon.variables.unitSpec = GetSpecialization()
-		if addon.variables.unitSpec then addon.variables.unitRole = GetSpecializationRole(GetSpecialization()) end
+		if addon.variables.unitSpec then
+			addon.variables.unitSpecName = select(2, GetSpecializationInfo(EnhanceQoL.variables.unitSpec))
+			addon.variables.unitRole = GetSpecializationRole(GetSpecialization())
+		end
 
 		if addon.db["showIlvlOnBagItems"] then
 			addon.functions.updateBags(ContainerFrameCombinedBags)
@@ -3738,7 +3741,10 @@ local eventHandlers = {
 	["PLAYER_LOGIN"] = function()
 		if addon.db["enableMinimapButtonBin"] then addon.functions.toggleButtonSink() end
 		addon.variables.unitSpec = GetSpecialization()
-		if addon.variables.unitSpec then addon.variables.unitRole = GetSpecializationRole(GetSpecialization()) end
+		if addon.variables.unitSpec then
+			addon.variables.unitSpecName = select(2, GetSpecializationInfo(EnhanceQoL.variables.unitSpec))
+			addon.variables.unitRole = GetSpecializationRole(GetSpecialization())
+		end
 
 		addon.db["moneyTracker"][UnitGUID("player")] = {
 			name = UnitName("player"),