+++ EnhanceQoL/Init.lua
@@ -493,8 +493,12 @@ addon.variables.unitClass = select(2, UnitClass("player"))
 addon.variables.unitClassID = select(3, UnitClass("player"))
 addon.variables.unitPlayerGUID = UnitGUID("player")
 addon.variables.unitSpec = GetSpecialization()
+addon.variables.unitSpecName = nil
 addon.variables.unitRole = nil
-if addon.variables.unitSpec then addon.variables.unitRole = GetSpecializationRole(GetSpecialization()) end
+if addon.variables.unitSpec then
+	addon.variables.unitSpecName = select(2, GetSpecializationInfo(EnhanceQoL.variables.unitSpec))
+	addon.variables.unitRole = GetSpecializationRole(GetSpecialization())
+end
 addon.variables.unitRace = select(2, UnitRace("player"))
 addon.variables.unitName = select(1, UnitName("player"))